backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
pydicom>=2.4.4
SimpleITK>=2.4.0
opencv-python-headless>=4.10.0
numpy>=2.0.0
scipy>=1.13.0
Pillow>=10.0.0
reportlab>=4.0.0
pytest>=8.0.0
httpx>=0.27.0
# Auth & billing
python-jose[cryptography]>=3.3.0
supabase>=2.0.0
stripe>=8.0.0
python-dotenv>=1.0.0
